requirements.txt
PyQt6
rawpy
Pillow
opencv-python
numpy
requests
exifread
